+++ core/src/tunnel.cpp
@@ -920,7 +920,8 @@ static std::shared_ptr<ServerUpstream> select_upstream(
             }
         }
         if (const SocketAddress *dst; // NOLINT(cppcoreguidelines-init-variables)
-                conn != nullptr && conn->flags.test(CONNF_LOOKINGUP_DOMAIN) && conn->flags.test(CONNF_SUSPECT_EXCLUSION)
+                conn != nullptr && conn->flags.test(CONNF_LOOKINGUP_DOMAIN)
+                && (conn->flags.test(CONNF_SUSPECT_EXCLUSION) || self->vpn->exclusions_tcp_early_ack)
                 && nullptr != (dst = std::get_if<SocketAddress>(&conn->addr.dst))
                 && is_domain_scannable_port(dst->port())) {
             return self->fake_upstream;