+++ core/src/vpn_client.cpp
@@ -262,6 +262,7 @@ VpnError VpnClient::init(const VpnSettings *settings) {
 
     this->tunnel->udp_close_wait_hostname_cache = g_udp_close_wait_hostname_cache;
     this->kill_switch_on = settings->killswitch_enabled;
+    this->exclusions_tcp_early_ack = settings->exclusions_tcp_early_ack;
     update_exclusions(settings->mode, {settings->exclusions.data, settings->exclusions.size});
 
     if (settings->tmp_files_base_path != nullptr) {